Refactor commit
