+++ src/neptune_exporter/storage/parquet_reader.py
@@ -18,16 +18,13 @@ import pyarrow as pa
 import pyarrow.parquet as pq
 import pyarrow.compute as pc
 from pathlib import Path
-from typing import Any, Generator, Optional, Union
+from typing import Any, Generator, Optional
 import logging
 from dataclasses import dataclass
 from neptune_exporter import model
-from neptune_exporter.cloud_storage import GCSPath
+from neptune_exporter.storage.gcs import GCSPath
 from neptune_exporter.types import ProjectId, RunFilePrefix, SourceRunId
 from neptune_exporter.utils import sanitize_path_part
-
-AnyPath = Union[Path, GCSPath]
-
 
 @dataclass(frozen=True, order=True, slots=True)
 class RunMetadata:
@@ -45,29 +42,9 @@ class RunMetadata:
 class ParquetReader:
     """Reads exported Neptune data from parquet files."""
 
-    def __init__(self, base_path: AnyPath):
+    def __init__(self, base_path: Path | GCSPath):
         self.base_path = base_path
         self._logger = logging.getLogger(__name__)
-        self._pa_filesystem = None  # lazily initialised for GCS paths
-
-    # ------------------------------------------------------------------
-    # Internal helpers
-    # ------------------------------------------------------------------
-
-    def _get_pa_filesystem(self):
-        """Return a PyArrow filesystem for GCS, or *None* for local paths."""
-        if isinstance(self.base_path, GCSPath):
-            if self._pa_filesystem is None:
-                self._pa_filesystem = self.base_path.get_pyarrow_filesystem()
-            return self._pa_filesystem
-        return None
-
-    def _read_parquet_table(self, path: AnyPath) -> pa.Table:
-        """Read a parquet file from either a local or GCS path."""
-        if isinstance(path, GCSPath):
-            pa_fs = self._get_pa_filesystem()
-            return pq.read_table(path.gcs_path, filesystem=pa_fs, schema=model.SCHEMA)
-        return pq.read_table(path, schema=model.SCHEMA)
 
     def check_run_exists(
         self,
@@ -101,7 +78,7 @@ class ParquetReader:
         self,
         project_ids: Optional[list[ProjectId]] = None,
         entity_scope: str | None = None,
-    ) -> list[AnyPath]:
+    ) -> list[Path | GCSPath]:
         """List all available projects in the exported data."""
         if not self.base_path.exists():
             return []
@@ -224,7 +201,13 @@ class ParquetReader:
 
         for part_file in run_part_files:
             try:
-                table = self._read_parquet_table(part_file)
+                if isinstance(part_file, GCSPath):
+                    pa_filesystem = part_file.get_pyarrow_filesystem()
+                    table = pq.read_table(part_file.gcs_path, filesystem=pa_filesystem, schema=model.SCHEMA)
+
+                else:
+                    table = pq.read_table(part_file, schema=model.SCHEMA)
+
                 table = self._filter_attributes(
                     table,
                     attribute_types=attribute_types,
